requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
fredapi>=0.5
yfinance>=0.2
python-dotenv>=1.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
jupyter>=1.0
pybind11>=2.11
setuptools>=68.0
